Revert .long() cast in Dequantize.inverse
The inverse is used during inference with compiled models where
tensors are already float; the original torch.floor(x) is correct.

+++ pointcountfm/preprocessing.py
@@ -159,7 +159,7 @@ class Dequantize(Transformation):
         return x.float() + torch.rand_like(x.float())
 
     def inverse(self, x: torch.Tensor) -> torch.Tensor:
-        return torch.floor(x).long()
+        return torch.floor(x)
 
 
 def compose(transformation: list[list[str | dict | list | None]] | None) -> Sequence:
